pyproject.toml
[tool.poetry]
name = "tabdpt"
version = "1.1.0"
description = "TabDPT"
authors = ["Jeremy", "Valentin"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "*"
numpy = "*"
scikit-learn = "*"
faiss-cpu = "*"
gdown = "*"
tqdm = "*"
scipy = "*"
omegaconf = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"